Command/scene_classification/Keras-VGG16-places365/requirements.txt
absl-py==0.9.0
aniso8601==8.1.1
argh==0.26.2
asn1crypto==0.24.0
astor==0.8.1
boto3==1.10.47
botocore==1.13.50
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
docker==4.4.2
docutils==0.15.2
english==2020.7.0
Flask==1.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
flask-url-map-serializer==0.0.1
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.35.0
grpcio-tools==1.35.0
h5py==2.10.0
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
MarkupSafe==1.1.1
monotonic==1.5
mpu==0.20.0
neobolt==1.7.17
neotime==1.7.4
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
packaging==20.9
pandas==1.0.3
pansi==2020.7.3
passlib==1.7.4
pika==1.1.0
Pillow==5.1.0
prompt-toolkit==2.0.10
protobuf==3.11.2
psycopg2-binary==2.8.6
py2neo==2020.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.8.0
pygobject==3.26.1
pymilvus==0.2.14
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pyxdg==0.25
PyYAML==5.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.22.1
scipy==1.4.1
SecretStorage==2.3.1
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
ujson==4.0.2
urllib3==1.24.3
watchdog==2.0.1
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
